refactor: rename Monaco editor component to a2ui-composer-monaco-editor and clean up redundant editor logic in raw-frame
